Add a facility to trigger sync updates externally
So far sync operations were executed either after a specific timeout or
if a database change has been detected. This commit adds an additional
facility for triggering sync updates externally. This is done with the
following request

`{ "RequestType": "sync" }\n`

Since sync services hold no data, the `Data` field of the response is
`null` but `ResponseType` will still indicate if the sync request was
committed successfully (`ok`) or not.

The python client and shell script have been update accordingly.

Signed-off-by: Spyros Stathopoulos <spystath@gmail.com>